[Dakshta/Keshav/Dhruvi] Refactor the code to build
